Add WiiStation-style perf baseline/comparison tooling; fix a 64-bit-to-32-bit truncation bug it exposed
Port of the parts of WiiStation's (sibling project, same machine)
profiling workflow that were missing here -- Wii64's own perf_prof.c
already says it was modeled on WiiStation's pattern, but only the
instrumentation calls existed, not the tooling to actually use them
for A/B comparison:

- scripts/sdimage_read.py: verbatim port, a generic FAT32 reader that
  walks a cluster chain directly. Needed because this project's
  Dolphin builds never reach a clean shutdown (see the shutdown-hang
  writeup in doc/subsystem-review.md), so perf.log's usual
  synced-folder copy is stale/missing -- this reads it straight off
  the raw SD image instead, which only needs the OS file lock to
  release (immediate on a plain kill).
- .dev/wii64_diag.sh: new `perf [out.log]` subcommand -- force-kills
  the tracked instance and extracts perf.log via sdimage_read.py in
  one step. Not committed (.dev/ is gitignored) but documented here
  since it's the actual entry point.
- .dev/build_profiling.sh: clean + rebuild with -DPERF_PROF in one
  step (Makefile.base has it commented out of DEBUG_FLAGS, and there's
  no header-dependency tracking, so a flag-only change needs the clean
  first). Also not committed, same reason.
- scripts/perf_compare.py: A/B comparison of two perf.log's averaged
  boxart page-load stats, adapted from WiiStation's script for Wii64's
  page/tile/mark log format instead of WiiStation's tagged counters.
- baselines/: WiiStation's baseline layout (index.csv + per-run
  perf.log + notes.md), scripts/baseline_add.py to file one. Filed
  2026-09-21_selectrom_baseline as the first real entry.

Getting real numbers flowing end-to-end (this was never actually
exercised before -- PERF_PROF has apparently never been enabled for a
build until today) surfaced a real bug: SelectRomFrame.cpp's
PERF_NOW()/PERF_US() timing locals were `unsigned int`, but
PERF_NOW() is gettime(), which returns a 64-bit tick count --
silently truncated to 32 bits, so every init/load/flush/total number
in perf.log was nonsense (order of 900+ seconds for a sub-millisecond
texture load). Fixed by declaring them `u64`. Before/after is exactly
the kind of regression perf_compare.py is for -- ran it on the two
captures as a first real test:
  avg_tile_load_us   954438167.8    991.6   0.000
Confirmed the fix with a fresh capture: sane sub-millisecond to
low-millisecond numbers now.

+++ menu/SelectRomFrame.cpp
@@ -719,7 +719,7 @@ void selectRomFrame_FillPage()
 	// someone deliberately opts in.
 	int boxartLoadLimit = BOXART_GetLoadLimit(NUM_FILE_SLOTS);
 #ifdef PERF_PROF
-	unsigned int pageStart = PERF_NOW();
+	u64 pageStart = PERF_NOW();
 #endif
 	perfProf_pageBegin(NUM_FILE_SLOTS, boxartLoadLimit);
 
@@ -743,12 +743,12 @@ void selectRomFrame_FillPage()
 				//			rom_headers[i+(current_page*NUM_FILE_SLOTS)].CRC1);
 				//load boxart
 #ifdef PERF_PROF
-				unsigned int t0 = PERF_NOW();
+				u64 t0 = PERF_NOW();
 #endif
 				BOXART_Init();
 #ifdef PERF_PROF
 				unsigned int initUs = PERF_US(t0);
-				unsigned int t1 = PERF_NOW();
+				u64 t1 = PERF_NOW();
 #endif
 #ifdef SHOW_DEBUG
 				bool found =
@@ -763,7 +763,7 @@ void selectRomFrame_FillPage()
 				}
 #endif
 #ifdef PERF_PROF
-				unsigned int t2 = PERF_NOW();
+				u64 t2 = PERF_NOW();
 #endif
 				DCFlushRange(fileTextures[i], BOXART_TEX_SIZE);
 				perfProf_tileLoaded(i, 1, initUs, loadUs, PERF_US(t2));
@@ -792,7 +792,7 @@ void selectRomFrame_FillPage()
 		}
 	}
 #ifdef PERF_PROF
-	unsigned int invalidateStart = PERF_NOW();
+	u64 invalidateStart = PERF_NOW();
 #endif
 	GX_InvalidateTexAll();
 	perfProf_pageEnd(PERF_US(invalidateStart), PERF_US(pageStart));
